+++ src/Electron2D/Core/InputEvent.cs
@@ -1,0 +1,5 @@
+namespace Electron2D;
+
+public class InputEvent : Resource
+{
+}